:recycle: Popular movies feed now inherits from a common feed
Next step is to make favorites feed inherit from it as well
